Change inet_ntop failure error code from EAFNOSUPPORT to EINVAL
EINVAL is the appropriate code for this. EAFNOSUPPORT is wrong because it
means a wrong address family was used, which is not the case here.
For functions that call bind(2)/connect(2) after inet_ntop this is in line
with bind/connect error codes: Both return EINVAL for an invalid address
argument as well.